requirements.txt
bcrypt
fastAPI
SqlAlchemy
pycryptodome
dotenv
uvicorn
python-jose